.js -> .tsx
switch to .tsx files to inherit react knowledge that we get from the practicum lessons
