+++ theatre/src/components/Dashboard.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useNavigate } from 'react-router-dom';
 import './Dashboard.css';
 
-function Dashboard() {
+const Dashboard: React.FC = () => {
   const navigate = useNavigate();
 
   const goToTheatreShows = () => {
@@ -17,12 +17,9 @@ function Dashboard() {
     <div className="dashboard-container">
       <h2>Admin Dashboard</h2>
       <div className="button-container">
-        {/* Button to go to Theatre Shows page */}
         <button onClick={goToTheatreShows} className="dashboard-button1">
           Theatre Shows
         </button>
-
-        {/* Button to go to Reservations page */}
         <button onClick={goToReservations} className="dashboard-button2">
           Reservations
         </button>
