+++ theatre/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import App from './App';
+import App from './App.tsx';
 import { BrowserRouter } from 'react-router-dom'; // Import BrowserRouter
 
 ReactDOM.render(
